[fix] revenue type: float -> int

+++ schema/revenue.py
@@ -3,4 +3,4 @@ from pydantic import BaseModel
 
 class Revenue(BaseModel):
     agg: float
-    value: float
+    value: int
